+++ app/src/main/kotlin/com/example/blogsearcher/app/search/BlogSearchEventHandler.kt
@@ -1,6 +1,8 @@
-package com.example.blogsearcher.app.rank
+package com.example.blogsearcher.app.search
 
 import com.example.blogsearcher.app.config.AsyncConfig
+import com.example.blogsearcher.domain.record.BlogSearchRecord
+import com.example.blogsearcher.domain.record.BlogSearchRecordRepository
 import com.example.blogsearcher.domain.search.event.BlogSearchEvent
 import com.github.benmanes.caffeine.cache.LoadingCache
 import org.springframework.context.event.EventListener
@@ -8,22 +10,21 @@ import org.springframework.scheduling.annotation.Async
 import org.springframework.stereotype.Component
 
 @Component
-class KeywordRankCalculator(
+class BlogSearchEventHandler(
+    private val blogSearchEventRepository: BlogSearchRecordRepository,
     private val cache: LoadingCache<String, Long>
 ) {
+    @Async(AsyncConfig.EVENT_POOL_NAME)
+    @EventListener(BlogSearchEvent::class)
+    fun handle(event: BlogSearchEvent) {
+        blogSearchEventRepository.save(BlogSearchRecord(event.vendor, event.keyword, event.searchAt))
+    }
+
     @Async(AsyncConfig.EVENT_POOL_NAME)
     @EventListener(BlogSearchEvent::class)
     fun saveOrIncrease(blogSearchEvent: BlogSearchEvent) {
         val keyword = blogSearchEvent.keyword.value
         val count = cache.get(keyword)
         cache.put(keyword, count + 1)
-    }
-
-    fun getTop(rank: Int): Map<String, Long> {
-        return cache.asMap()
-            .toList()
-            .sortedByDescending { it.second }
-            .take(rank)
-            .associate { it.first to it.second }
     }
 }
